feat/split-e2e-tests: splitted tests for faster execution
