+++ apps/chat-e2e/src/tests/chatBarConversation.test.ts
@@ -55,227 +55,6 @@ dialTest(
 );
 
 dialTest(
-  'Rename chat. Cancel.\n' +
-    'Long Chat name is cut. Named automatically by the system.\n' +
-    'Long chat name while delete and edit',
-  async ({
-    dialHomePage,
-    conversations,
-    conversationDropdownMenu,
-    conversationData,
-    localStorageManager,
-    dataInjector,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-588', 'EPMRTC-816', 'EPMRTC-1494');
-    const newName = 'new name to cancel';
-    let conversation: Conversation;
-    const conversationName = GeneratorUtil.randomString(70);
-
-    await dialTest.step('Prepare conversation with long name', async () => {
-      conversation = conversationData.prepareDefaultConversation(
-        gpt35Model,
-        conversationName,
-      );
-      await dataInjector.createConversations([conversation]);
-      await localStorageManager.setSelectedConversation(conversation);
-    });
-
-    await dialTest.step(
-      'Open app and verify conversation name is truncated in the side panel',
-      async () => {
-        await dialHomePage.openHomePage();
-        await dialHomePage.waitForPageLoaded();
-        const chatNameOverflow = await conversations
-          .getConversationName(conversationName)
-          .getComputedStyleProperty(Styles.text_overflow);
-        expect
-          .soft(chatNameOverflow[0], ExpectedMessages.chatNameIsTruncated)
-          .toBe(Overflow.ellipsis);
-      },
-    );
-
-    await dialTest.step(
-      'Hover over conversation name and verify it is truncated when menu dots appear',
-      async () => {
-        await conversations.getConversationByName(conversationName).hover();
-        const chatNameOverflow = await conversations
-          .getConversationName(conversationName)
-          .getComputedStyleProperty(Styles.text_overflow);
-        expect
-          .soft(chatNameOverflow[0], ExpectedMessages.chatNameIsTruncated)
-          .toBe(Overflow.ellipsis);
-      },
-    );
-
-    await dialTest.step(
-      'Select "Rename" from chat menu and verify it is truncated, cursor set at the end',
-      async () => {
-        await conversations.openConversationDropdownMenu(conversationName);
-        await conversationDropdownMenu.selectMenuOption(MenuOptions.rename);
-        const chatNameOverflow = await conversations
-          .getConversationName(conversationName)
-          .getComputedStyleProperty(Styles.text_overflow);
-        expect
-          .soft(chatNameOverflow[0], ExpectedMessages.chatNameIsTruncated)
-          .toBe(undefined);
-      },
-    );
-
-    await dialTest.step(
-      'Set new conversation name, cancel edit and verify conversation with initial name shown',
-      async () => {
-        const nameInput = await conversations.openEditConversationNameMode(
-          conversation.name,
-          newName,
-        );
-        await nameInput.clickCancelButton();
-        expect
-          .soft(
-            await conversations.getConversationByName(newName).isVisible(),
-            ExpectedMessages.conversationNameNotUpdated,
-          )
-          .toBeFalsy();
-      },
-    );
-
-    await dialTest.step(
-      'Select "Delete" from conversation menu and verify its name is truncated when menu dots appear',
-      async () => {
-        await conversations.openConversationDropdownMenu(conversationName);
-        await conversationDropdownMenu.selectMenuOption(MenuOptions.delete);
-        const chatNameOverflow = await conversations
-          .getConversationName(conversationName)
-          .getComputedStyleProperty(Styles.text_overflow);
-        expect
-          .soft(chatNameOverflow[0], ExpectedMessages.chatNameIsTruncated)
-          .toBe(Overflow.ellipsis);
-      },
-    );
-  },
-);
-
-dialTest(
-  'Rename chat before starting the conversation.\n' +
-    'Long Chat name is cut. Named manually',
-  async ({
-    dialHomePage,
-    conversations,
-    conversationDropdownMenu,
-    chat,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-584', 'EPMRTC-819');
-    const newName = GeneratorUtil.randomString(70);
-    await dialHomePage.openHomePage();
-    await dialHomePage.waitForPageLoaded({ isNewConversationVisible: true });
-    await conversations.openConversationDropdownMenu(
-      ExpectedConstants.newConversationTitle,
-    );
-    await conversationDropdownMenu.selectMenuOption(MenuOptions.rename);
-    await conversations.editConversationNameWithTick(
-      ExpectedConstants.newConversationTitle,
-      newName,
-    );
-    expect
-      .soft(
-        await conversations.getConversationByName(newName).isVisible(),
-        ExpectedMessages.conversationNameUpdated,
-      )
-      .toBeTruthy();
-
-    const chatNameOverflow = await conversations
-      .getConversationName(newName)
-      .getComputedStyleProperty(Styles.text_overflow);
-    expect
-      .soft(chatNameOverflow[0], ExpectedMessages.chatNameIsTruncated)
-      .toBe(Overflow.ellipsis);
-
-    await chat.sendRequestWithButton('one more test message');
-    expect
-      .soft(
-        await conversations.getConversationByName(newName).isVisible(),
-        ExpectedMessages.conversationNameUpdated,
-      )
-      .toBeTruthy();
-  },
-);
-
-dialTest(
-  'Rename chat after starting the conversation.\n' +
-    'Long Chat name is cut in chat header. Named manually.\n' +
-    'Tooltip shows full long chat name in chat header. Named manually.\n' +
-    'Long chat name is cut in chat header. Named automatically by the system.\n' +
-    'Tooltip shows full long chat name in chat header. Named automatically by the system',
-  async ({
-    dialHomePage,
-    conversations,
-    conversationDropdownMenu,
-    conversationData,
-    dataInjector,
-    localStorageManager,
-    chatHeader,
-    tooltip,
-    setTestIds,
-    errorPopup,
-  }) => {
-    setTestIds(
-      'EPMRTC-585',
-      'EPMRTC-821',
-      'EPMRTC-822',
-      'EPMRTC-818',
-      'EPMRTC-820',
-    );
-    const newName = GeneratorUtil.randomString(60);
-    const conversation = conversationData.prepareDefaultConversation();
-    await dataInjector.createConversations([conversation]);
-    await localStorageManager.setSelectedConversation(conversation);
-
-    await dialHomePage.openHomePage();
-    await dialHomePage.waitForPageLoaded();
-    await conversations.openConversationDropdownMenu(conversation.name);
-    await conversationDropdownMenu.selectMenuOption(MenuOptions.rename);
-    await conversations.editConversationNameWithEnter(
-      conversation.name,
-      newName,
-    );
-    expect
-      .soft(
-        await conversations.getConversationByName(newName).isVisible(),
-        ExpectedMessages.conversationNameUpdated,
-      )
-      .toBeTruthy();
-
-    const isChatHeaderTitleTruncated =
-      await chatHeader.chatTitle.isElementWidthTruncated();
-    expect
-      .soft(
-        isChatHeaderTitleTruncated,
-        ExpectedMessages.chatHeaderTitleTruncated,
-      )
-      .toBeTruthy();
-    await errorPopup.cancelPopup();
-    await chatHeader.chatTitle.hoverOver();
-    const tooltipChatHeaderTitle = await tooltip.getContent();
-    expect
-      .soft(
-        tooltipChatHeaderTitle,
-        ExpectedMessages.headerTitleCorrespondRequest,
-      )
-      .toBe(newName);
-
-    const isTooltipChatHeaderTitleTruncated =
-      await tooltip.isElementWidthTruncated();
-    expect
-      .soft(
-        isTooltipChatHeaderTitleTruncated,
-        ExpectedMessages.headerTitleIsFullyVisible,
-      )
-      .toBeFalsy();
-  },
-);
-
-dialTest(
   'Menu for New conversation',
   async ({
     dialHomePage,
@@ -341,71 +120,6 @@ dialTest(
         MenuOptions.publish,
         MenuOptions.delete,
       ]);
-  },
-);
-
-dialTest(
-  'Delete chat in the folder',
-  async ({
-    dialHomePage,
-    folderConversations,
-    conversationData,
-    dataInjector,
-    conversationDropdownMenu,
-    setTestIds,
-    confirmationDialog,
-  }) => {
-    setTestIds('EPMRTC-607');
-    const conversationInFolder =
-      conversationData.prepareDefaultConversationInFolder();
-    await dataInjector.createConversations(
-      conversationInFolder.conversations,
-      conversationInFolder.folders,
-    );
-
-    await dialHomePage.openHomePage();
-    await dialHomePage.waitForPageLoaded();
-    await folderConversations.expandFolder(conversationInFolder.folders.name);
-    await folderConversations.openFolderEntityDropdownMenu(
-      conversationInFolder.folders.name,
-      conversationInFolder.conversations[0].name,
-    );
-    await conversationDropdownMenu.selectMenuOption(MenuOptions.delete);
-    await confirmationDialog.confirm({ triggeredHttpMethod: 'DELETE' });
-    await folderConversations
-      .getFolderEntity(
-        conversationInFolder.folders.name,
-        conversationInFolder.conversations[0].name,
-      )
-      .waitFor({ state: 'hidden' });
-  },
-);
-
-dialTest(
-  'Delete chat located in the root',
-  async ({
-    dialHomePage,
-    conversations,
-    conversationDropdownMenu,
-    conversationData,
-    dataInjector,
-    localStorageManager,
-    confirmationDialog,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-608');
-    const conversation = conversationData.prepareDefaultConversation();
-    await dataInjector.createConversations([conversation]);
-    await localStorageManager.setSelectedConversation(conversation);
-
-    await dialHomePage.openHomePage({ iconsToBeLoaded: [gpt35Model.iconUrl] });
-    await dialHomePage.waitForPageLoaded();
-    await conversations.openConversationDropdownMenu(conversation.name);
-    await conversationDropdownMenu.selectMenuOption(MenuOptions.delete);
-    await confirmationDialog.confirm({ triggeredHttpMethod: 'DELETE' });
-    await conversations
-      .getConversationByName(conversation.name)
-      .waitFor({ state: 'hidden' });
   },
 );
 
@@ -507,342 +221,6 @@ dialTest.skip(
         ExpectedMessages.conversationOfToday,
       )
       .toBeTruthy();
-  },
-);
-
-dialTest(
-  'Chat is moved to folder created from Move to',
-  async ({
-    dialHomePage,
-    conversations,
-    conversationDropdownMenu,
-    conversationData,
-    dataInjector,
-    localStorageManager,
-    folderConversations,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-864');
-    const conversation = conversationData.prepareDefaultConversation();
-    await dataInjector.createConversations([conversation]);
-    await localStorageManager.setSelectedConversation(conversation);
-
-    await dialHomePage.openHomePage();
-    await dialHomePage.waitForPageLoaded();
-    await conversations.openConversationDropdownMenu(conversation.name);
-    await conversationDropdownMenu.selectMenuOption(MenuOptions.moveTo);
-    await conversations.selectMoveToMenuOption(
-      ExpectedConstants.newFolderTitle,
-    );
-
-    await folderConversations.expandFolder(
-      ExpectedConstants.newFolderWithIndexTitle(1),
-    );
-    const isFolderConversationVisible =
-      await folderConversations.isFolderEntityVisible(
-        ExpectedConstants.newFolderWithIndexTitle(1),
-        conversation.name,
-      );
-    expect
-      .soft(
-        isFolderConversationVisible,
-        ExpectedMessages.conversationMovedToFolder,
-      )
-      .toBeTruthy();
-
-    const folderNameColor = await folderConversations.getFolderNameColor(
-      ExpectedConstants.newFolderWithIndexTitle(1),
-    );
-    expect
-      .soft(folderNameColor[0], ExpectedMessages.folderNameColorIsValid)
-      .toBe(Colors.textAccentSecondary);
-  },
-);
-
-dialTest(
-  'Chat is moved to folder from Move to list.\n' +
-    'Long folder name is cut in Move to menu',
-  async ({
-    dialHomePage,
-    conversations,
-    conversationDropdownMenu,
-    conversationData,
-    localStorageManager,
-    dataInjector,
-    folderConversations,
-    folderDropdownMenu,
-    chatBar,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-863', 'EPMRTC-942');
-    const folderName = GeneratorUtil.randomString(70);
-    let conversation: Conversation;
-
-    await dialTest.step(
-      'Prepare conversation and folder with long name to move conversation in',
-      async () => {
-        conversation = conversationData.prepareDefaultConversation();
-        await dataInjector.createConversations([conversation]);
-        await localStorageManager.setSelectedConversation(conversation);
-      },
-    );
-
-    await dialTest.step(
-      'Open "Move to" menu option for conversation and verify folder name is truncated',
-      async () => {
-        await dialHomePage.openHomePage();
-        await dialHomePage.waitForPageLoaded();
-        await chatBar.createNewFolder();
-        await folderConversations.openFolderDropdownMenu(
-          ExpectedConstants.newFolderWithIndexTitle(1),
-          1,
-        );
-        await folderDropdownMenu.selectMenuOption(MenuOptions.rename);
-        await folderConversations.editFolderNameWithEnter(
-          ExpectedConstants.newFolderWithIndexTitle(1),
-          folderName,
-        );
-
-        await conversations.openConversationDropdownMenu(conversation.name);
-        await conversationDropdownMenu.selectMenuOption(MenuOptions.moveTo);
-
-        const moveToFolder =
-          await conversationDropdownMenu.getMenuOption(folderName);
-        await moveToFolder.waitForState();
-        const moveToFolderOverflow =
-          await moveToFolder.getComputedStyleProperty(Styles.text_overflow);
-        expect
-          .soft(moveToFolderOverflow[0], ExpectedMessages.folderNameIsTruncated)
-          .toBe(Overflow.ellipsis);
-      },
-    );
-
-    await dialTest.step(
-      'Select folder name from menu and conversation is moved into folder',
-      async () => {
-        await conversations.selectMoveToMenuOption(folderName);
-        await folderConversations.expandFolder(folderName);
-        const isFolderConversationVisible =
-          await folderConversations.isFolderEntityVisible(
-            folderName,
-            conversation.name,
-          );
-        expect
-          .soft(
-            isFolderConversationVisible,
-            ExpectedMessages.conversationMovedToFolder,
-          )
-          .toBeTruthy();
-      },
-    );
-  },
-);
-
-dialTest(
-  'Delete all conversations. Cancel',
-  async ({
-    dialHomePage,
-    conversations,
-    conversationData,
-    folderConversations,
-    chatBar,
-    confirmationDialog,
-    dataInjector,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-610');
-    conversationData.resetData();
-    const singleConversation = conversationData.prepareDefaultConversation();
-    conversationData.resetData();
-    const conversationInFolder =
-      conversationData.prepareDefaultConversationInFolder();
-    await dataInjector.createConversations(
-      [singleConversation, ...conversationInFolder.conversations],
-      conversationInFolder.folders,
-    );
-
-    await dialHomePage.openHomePage();
-    await dialHomePage.waitForPageLoaded();
-    await chatBar.createNewFolder();
-    await folderConversations.expandFolder(conversationInFolder.folders.name);
-    await chatBar.deleteAllEntities();
-    await confirmationDialog.cancelDialog();
-
-    const isFolderConversationVisible =
-      await folderConversations.isFolderEntityVisible(
-        conversationInFolder.folders.name,
-        conversationInFolder.conversations[0].name,
-      );
-    expect
-      .soft(
-        isFolderConversationVisible,
-        ExpectedMessages.conversationNotDeleted,
-      )
-      .toBeTruthy();
-
-    const isFolderVisible = await folderConversations
-      .getFolderByName(ExpectedConstants.newFolderWithIndexTitle(1))
-      .isVisible();
-    expect
-      .soft(isFolderVisible, ExpectedMessages.folderNotDeleted)
-      .toBeTruthy();
-
-    const isSingleConversationVisible = await conversations
-      .getConversationByName(singleConversation.name)
-      .isVisible();
-    expect
-      .soft(
-        isSingleConversationVisible,
-        ExpectedMessages.conversationNotDeleted,
-      )
-      .toBeTruthy();
-  },
-);
-
-dialTest(
-  'Delete all conversations. Clear',
-  async ({
-    dialHomePage,
-    conversations,
-    conversationData,
-    promptData,
-    dataInjector,
-    folderConversations,
-    chatBar,
-    confirmationDialog,
-    folderPrompts,
-    prompts,
-    promptBar,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-611');
-    const singleConversation = conversationData.prepareDefaultConversation();
-    conversationData.resetData();
-    const conversationInFolder =
-      conversationData.prepareDefaultConversationInFolder();
-    conversationData.resetData();
-    const promptInFolder = promptData.prepareDefaultPromptInFolder();
-    promptData.resetData();
-    const singlePrompt = promptData.prepareDefaultPrompt();
-
-    await dialHomePage.openHomePage();
-    await dialHomePage.waitForPageLoaded();
-    if (isApiStorageType) {
-      await dataInjector.createConversations([
-        singleConversation,
-        ...conversationInFolder.conversations,
-      ]);
-      await dataInjector.createPrompts([
-        singlePrompt,
-        ...promptInFolder.prompts,
-      ]);
-    } else {
-      await dataInjector.updateConversations(
-        [singleConversation, ...conversationInFolder.conversations],
-        conversationInFolder.folders,
-      );
-      await dataInjector.updatePrompts(
-        [singlePrompt, ...promptInFolder.prompts],
-        promptInFolder.folders,
-      );
-    }
-
-    await dialHomePage.reloadPage();
-    await dialHomePage.waitForPageLoaded();
-    await promptBar.createNewFolder();
-    for (let i = 1; i <= 4; i++) {
-      await chatBar.createNewFolder();
-    }
-    for (let i = 3; i >= 2; i--) {
-      await chatBar.dragAndDropEntityToFolder(
-        folderConversations.getFolderByName(
-          ExpectedConstants.newFolderWithIndexTitle(i),
-        ),
-        folderConversations.getFolderByName(
-          ExpectedConstants.newFolderWithIndexTitle(i - 1),
-        ),
-      );
-    }
-    await folderConversations.expandFolder(
-      ExpectedConstants.newFolderWithIndexTitle(2),
-    );
-
-    await folderConversations.expandFolder(conversationInFolder.folders.name);
-    await folderPrompts.expandFolder(promptInFolder.folders.name);
-    await chatBar.deleteAllEntities();
-    await confirmationDialog.confirm({ triggeredHttpMethod: 'DELETE' });
-
-    let i = 2;
-    while (i > 0) {
-      const isFolderConversationVisible =
-        await folderConversations.isFolderEntityVisible(
-          conversationInFolder.folders.name,
-          conversationInFolder.conversations[0].name,
-        );
-      expect
-        .soft(isFolderConversationVisible, ExpectedMessages.conversationDeleted)
-        .toBeFalsy();
-
-      const isFolderVisible = await folderConversations
-        .getFolderByName(ExpectedConstants.newFolderWithIndexTitle(4))
-        .isVisible();
-      expect.soft(isFolderVisible, ExpectedMessages.folderDeleted).toBeFalsy();
-
-      for (let i = 1; i <= 3; i++) {
-        const isNestedFolderVisible = await folderConversations
-          .getFolderByName(ExpectedConstants.newFolderWithIndexTitle(i))
-          .isVisible();
-        expect
-          .soft(isNestedFolderVisible, ExpectedMessages.folderDeleted)
-          .toBeFalsy();
-      }
-
-      const isSingleConversationVisible = await conversations
-        .getConversationByName(singleConversation.name)
-        .isVisible();
-      expect
-        .soft(isSingleConversationVisible, ExpectedMessages.conversationDeleted)
-        .toBeFalsy();
-
-      await conversations
-        .getConversationByName(ExpectedConstants.newConversationTitle)
-        .waitFor();
-
-      if (i === 1) {
-        await folderPrompts.expandFolder(promptInFolder.folders.name);
-      }
-      const isFolderPromptVisible = await folderPrompts.isFolderEntityVisible(
-        promptInFolder.folders.name,
-        promptInFolder.prompts[0].name,
-      );
-      expect
-        .soft(isFolderPromptVisible, ExpectedMessages.promptNotDeleted)
-        .toBeTruthy();
-
-      const isPromptFolderVisible = await folderPrompts
-        .getFolderByName(ExpectedConstants.newFolderWithIndexTitle(1))
-        .isVisible();
-      i === 1
-        ? expect
-            .soft(isPromptFolderVisible, ExpectedMessages.folderNotDeleted)
-            .toBeFalsy()
-        : expect
-            .soft(isPromptFolderVisible, ExpectedMessages.folderNotDeleted)
-            .toBeTruthy();
-
-      const isSinglePromptVisible = await prompts
-        .getPromptByName(singlePrompt.name)
-        .isVisible();
-      expect
-        .soft(isSinglePromptVisible, ExpectedMessages.promptNotDeleted)
-        .toBeTruthy();
-
-      if (i > 1) {
-        await dialHomePage.reloadPage();
-        await dialHomePage.waitForPageLoaded();
-      }
-      i--;
-    }
   },
 );
