+++ apps/chat-e2e/src/tests/replay.test.ts
@@ -399,79 +399,6 @@ dialTest(
 );
 
 dialTest(
-  'Restart replay after error appeared on browser refresh.\n' +
-    'Restart replay after error appeared on network interruption',
-  async ({
-    dialHomePage,
-    conversationData,
-    chat,
-    localStorageManager,
-    dataInjector,
-    setTestIds,
-    chatMessages,
-    tooltip,
-    context,
-  }) => {
-    setTestIds('EPMRTC-514', 'EPMRTC-1165');
-    let conversation: Conversation;
-    let replayConversation: Conversation;
-    await dialTest.step('Prepare conversation to replay', async () => {
-      conversation = conversationData.prepareDefaultConversation(gpt35Model);
-      replayConversation =
-        conversationData.prepareDefaultReplayConversation(conversation);
-      await dataInjector.createConversations([
-        conversation,
-        replayConversation,
-      ]);
-      await localStorageManager.setSelectedConversation(replayConversation);
-    });
-
-    await dialTest.step(
-      'Press Start replay and interrupt it with network error',
-      async () => {
-        await dialHomePage.openHomePage();
-        await dialHomePage.waitForPageLoaded();
-        await context.setOffline(true);
-        await chat.startReplay();
-      },
-    );
-
-    await dialTest.step('Verify error message is displayed', async () => {
-      const generatedContent = await chatMessages.getLastMessageContent();
-
-      await chat.proceedGenerating.hoverOver();
-      const tooltipContent = await tooltip.getContent();
-
-      expect
-        .soft(generatedContent, ExpectedMessages.errorReceivedOnReplay)
-        .toBe(ExpectedConstants.answerError);
-      expect
-        .soft(tooltipContent, ExpectedMessages.proceedReplayIsVisible)
-        .toBe(ExpectedConstants.continueReplayAfterErrorLabel);
-    });
-
-    await dialTest.step(
-      'Proceed replaying and verify response received',
-      async () => {
-        await context.setOffline(false);
-        await chat.proceedReplaying(true);
-        const generatedContent = await chatMessages.getGeneratedChatContent(
-          conversation.messages.length,
-        );
-        expect
-          .soft(
-            generatedContent.includes(
-              conversation.messages.find((m) => m.role === 'user')!.content,
-            ),
-            ExpectedMessages.replayContinuesFromReceivedContent,
-          )
-          .toBeTruthy();
-      },
-    );
-  },
-);
-
-dialTest(
   '"Replay as is" when chat is based on Model.\n' +
     '"Replay as is" when chat is based on Model with addon',
   async ({
@@ -764,7 +691,7 @@ dialTest(
 
 dialTest(
   'Replay function is still available if the name was edited.\n' +
-    'Start replay works in  renamed [Replay]chat.\n' +
+    'Start replay works in renamed [Replay]chat.\n' +
     'Regenerate response in already replayed chat.\n' +
     'Continue conversation in already replayed chat',
   async ({
@@ -850,118 +777,6 @@ dialTest(
             ExpectedMessages.chatRequestMessageIsValid,
           )
           .toBe(newMessage);
-      },
-    );
-  },
-);
-
-dialTest(
-  'Start replay button appears in [Replay]chat if the parent chat has error in the response',
-  async ({
-    dialHomePage,
-    conversationData,
-    chat,
-    localStorageManager,
-    dataInjector,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-1312');
-    let errorConversation: Conversation;
-    let replayConversation: Conversation;
-
-    await dialTest.step(
-      'Prepare errorConversation with error response and replay errorConversation',
-      async () => {
-        errorConversation =
-          conversationData.prepareErrorResponseConversation(gpt35Model);
-        replayConversation =
-          conversationData.prepareDefaultReplayConversation(errorConversation);
-        await dataInjector.createConversations([
-          errorConversation,
-          replayConversation,
-        ]);
-        await localStorageManager.setSelectedConversation(replayConversation);
-      },
-    );
-
-    await dialTest.step(
-      'Verify "Start Replay" button is available',
-      async () => {
-        await dialHomePage.openHomePage();
-        await dialHomePage.waitForPageLoaded();
-
-        const isStartReplayEnabled = await chat.replay.isElementEnabled();
-        expect
-          .soft(isStartReplayEnabled, ExpectedMessages.startReplayVisible)
-          .toBeTruthy();
-      },
-    );
-  },
-);
-
-dialTest(
-  `"Replay as is" when restricted Model is used in parent chat`,
-  async ({
-    dialHomePage,
-    conversationData,
-    chat,
-    localStorageManager,
-    dataInjector,
-    talkToSelector,
-    setTestIds,
-  }) => {
-    setTestIds('EPMRTC-1328');
-    let notAllowedModelConversation: Conversation;
-    let replayConversation: Conversation;
-
-    await dialTest.step(
-      'Prepare conversation with not allowed model and replay for it',
-      async () => {
-        notAllowedModelConversation =
-          conversationData.prepareDefaultConversation('not_allowed_model');
-        replayConversation = conversationData.prepareDefaultReplayConversation(
-          notAllowedModelConversation,
-        );
-        await dataInjector.createConversations([
-          notAllowedModelConversation,
-          replayConversation,
-        ]);
-        await localStorageManager.setSelectedConversation(replayConversation);
-      },
-    );
-
-    await dialTest.step(
-      'Verify "Start Replay" button is not displayed, error is shown at the bottom',
-      async () => {
-        await dialHomePage.openHomePage();
-        await dialHomePage.waitForPageLoaded();
-
-        await talkToSelector.waitForState({ state: 'attached' });
-
-        const isStartReplayVisible = await chat.replay.isVisible();
-        expect
-          .soft(isStartReplayVisible, ExpectedMessages.startReplayNotVisible)
-          .toBeFalsy();
-
-        const notAllowedModelError =
-          await chat.notAllowedModelLabel.getElementContent();
-        expect
-          .soft(
-            notAllowedModelError!.trim(),
-            ExpectedMessages.notAllowedModelErrorDisplayed,
-          )
-          .toBe(ExpectedConstants.notAllowedModelError);
-      },
-    );
-
-    await dialTest.step(
-      'Select any available model and start replaying',
-      async () => {
-        await talkToSelector.selectModel(gpt35Model);
-        const replayRequest = await chat.startReplay();
-        expect
-          .soft(replayRequest.modelId, ExpectedMessages.chatRequestModelIsValid)
-          .toBe(gpt35Model.id);
       },
     );
   },
